fix(util): replace lodash deps and remove unused utils
