+++ packages/admin-ui-util/src/__tests__/string.test.ts
@@ -1,4 +1,19 @@
-import { isKebab } from '../isKebab'
+import { isKebab, capitalize } from '../string'
+
+describe('capitalize', () => {
+  it('should capitalize a string', () => {
+    expect(capitalize('')).toBe('')
+    expect(capitalize('text')).toBe('Text')
+  })
+
+  it('should not cause side effects', () => {
+    const value = 'text'
+    const capitalized = capitalize(value)
+
+    expect(value).toBe('text')
+    expect(capitalized).toBe('Text')
+  })
+})
 
 describe('isKebab', () => {
   it('should be true if is kebab case', () => {
