+++ packages/admin-ui-util/src/string.ts
@@ -1,4 +1,14 @@
 /**
+ * Capitalizes a string
+ * @param str
+ * @example
+ * capitalize('string') // 'String'
+ */
+export function capitalize(str: string) {
+  return str.charAt(0).toUpperCase() + str.slice(1)
+}
+
+/**
  * Check whether the string is in kebab case
  * @param name
  * @returns boolean
