+++ packages/admin-ui/src/components/RadioGroup/index.tsx
@@ -2,7 +2,6 @@ import type { ReactNode } from 'react'
 import React, { Fragment } from 'react'
 import type { RadioGroupProps as ReakitRadioGroupProps } from 'reakit/Radio'
 import { RadioGroup as ReakitRadioGroup } from 'reakit/Radio'
-import type { StyleObject } from '@vtex/admin-ui-core'
 import { tag } from '@vtex/admin-ui-react'
 import { get } from '@vtex/admin-ui-util'
 
@@ -51,7 +50,7 @@ export function RadioGroup(props: RadioGroupProps) {
             display: 'flex',
             alignItems: 'center',
           },
-          ...get<Record<string, StyleObject>, string, object>(
+          ...get(
             {
               'horizontal-regular': {
                 flexDirection: 'row',
