--- packages/util/src/index.ts
@@ -1,17 +1,0 @@
-export * from './lodash'
-export * from './isObjectEmpty'
-export * from './isFunction'
-export * from './flatMap'
-export * from './hasDuplicates'
-export * from './callOrReturn'
-export * from './capitalize'
-export * from './isBrowser'
-export * from './isKebab'
-export * from './forwardRef'
-export * from './renameKeys'
-export * from './types'
-export * from './callAllHandlers'
-export * from './attr'
-export * from './isTouch'
-export * from './focus'
-export * from './runIfFunction'
