--- packages/util/src/isBrowser.ts
@@ -1,9 +1,0 @@
-export function canUseDOM(): boolean {
-  return !!(
-    typeof window !== 'undefined' &&
-    window.document &&
-    window.document.createElement
-  )
-}
-
-export const isBrowser = canUseDOM()
