--- packages/util/src/isFunction.ts
@@ -1,8 +1,0 @@
-/**
- * Whether a value is a function
- */
-export function isFunction<T extends Function = Function>(
-  value: any
-): value is T {
-  return typeof value === 'function'
-}
