--- packages/util/src/lodash.ts
@@ -1,8 +1,0 @@
-import get from 'lodash.get'
-import omit from 'lodash.omit'
-import pick from 'lodash.pick'
-import lodashMerge from 'lodash.merge'
-
-const merge = (...params: any) => lodashMerge({}, ...params)
-
-export { get, omit, pick, merge }
