--- packages/util/src/runIfFunction.ts
@@ -1,8 +1,0 @@
-import { isFunction } from './isFunction'
-
-export function runIfFunction<T, U>(
-  valueOrFn: T | ((...fnArgs: U[]) => T),
-  ...args: U[]
-): T {
-  return isFunction(valueOrFn) ? valueOrFn(...args) : valueOrFn
-}
